# book-network/src/main/java/com/hab/book/book/BorrowedBookResponse.java
package com.hab.book.book;

import lombok.Builder;

@Builder
public record BorrowedBookResponse(
        Integer id,
        String title,
        String authorName,
        String isbn,
        double rate,
        boolean returned,
        boolean returnedApproved
) {
}